提交代码

+++ database/migrations/2019_08_21_112835_create_wechat_users_table.php
@@ -15,7 +15,7 @@ class CreateWechatUsersTable extends Migration
     {
         Schema::create('wechat_users', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('uid');
+            $table->integer('uid');
             $table->string('openid');
             $table->string('unionid');
             $table->string('type')->comment('DYH,FWH,MP');
